requirements.txt
botcity-framework-core>=1.0.0,<2.0
pandas==1.5.1
openpyxl
